+++ src/newtypes.rs
@@ -101,3 +101,8 @@ newtype_f64!(Percent, "percent", "%");
 newtype_f64!(Picometer, "picometers", "pm");
 newtype_f64!(KiloJoulePerMol, "kilojoules per mol", "kJ/mol");
 newtype_f64!(Kelvin, "Kelvin", "K");
+newtype_f64!(
+    GramPerCubicCentimeter,
+    "grams per cubic centimeter",
+    "g/cm³"
+);