+++ src/properties.rs
@@ -36,3 +36,4 @@ prop!(electronic_configuration, "electronic_configuration");
 prop!(discoverers, "discoverers");
 prop!(discovery_location, "discovery_location");
 prop!(oxidation_states, "oxidation_states");
+prop!(density, "density");