+++ src/properties.rs
@@ -37,3 +37,4 @@ prop!(discoverers, "discoverers");
 prop!(discovery_location, "discovery_location");
 prop!(oxidation_states, "oxidation_states");
 prop!(density, "density");
+prop!(electron_affinity, "electron_affinity");