+++ src/properties.rs
@@ -38,3 +38,4 @@ prop!(discovery_location, "discovery_location");
 prop!(oxidation_states, "oxidation_states");
 prop!(density, "density");
 prop!(electron_affinity, "electron_affinity");
+prop!(ionization_energy, "ionization_energy");