+++ src/properties.rs
@@ -36,3 +36,6 @@ prop!(electronic_configuration, "electronic_configuration");
 prop!(discoverers, "discoverers");
 prop!(discovery_location, "discovery_location");
 prop!(oxidation_states, "oxidation_states");
+prop!(density, "density");
+prop!(electron_affinity, "electron_affinity");
+prop!(ionization_energy, "ionization_energy");