+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - `Element::atomic_radius()` return type changed to `Option<Picometer>`
+- `Element::melting_point()` return type changed to `Option<Kelvin>`
+- `Element::boiling_point()` return type changed to `Option<Kelvin>`
 
 ## [0.7.0] - 2023-02-06